requirements.txt
absl-py==0.11.0
aiohttp==4.0.0a1
aiohttp-cors==0.7.0
alembic==1.6.5
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
audioread==2.1.9
bandmat==0.7
BentoML==0.13.0
boto3==1.17.99
botocore==1.20.99
Bottleneck==1.3.2
cached-property==1.5.2
cachetools==4.1.1
Cerberus==1.3.4
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
configparser==5.0.1
contextlib2==0.6.0.post1
contextvars==2.4
ctc-segmentation==1.6.5
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
deepmerge==0.3.0
dependency-injector==4.33.0
Distance==0.1.3
docker==5.0.0
docker-pycreds==0.4.0
editdistance==0.5.2
-e git+https://github.com/qdenisq/espnet.git@78f5e942c1c6aac1341b319993c2292dfa371c82#egg=espnet
espnet-model-zoo==0.0.0a20
espnet-tts-frontend==0.0.3
fastdtw==0.3.4
filelock==3.0.12
Flask==1.1.2
future==0.18.2
g2p-en==2.1.0
gdown==3.12.2
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
grpcio==1.33.2
gunicorn==20.1.0
h5py==3.1.0
humanfriendly==9.0
idna==2.10
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==3.1.0
inflect==5.0.2
isort==5.6.4
itsdangerous==1.1.0
jaconv==0.2.4
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
kaldiio==2.17.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
librosa==0.8.0
llvmlite==0.34.0
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mir-eval==0.6
Morfessor==2.0.6
multidict==4.7.6
musdb==0.3.1
museval==0.3.1
nara-wpe==0.0.7
nltk==3.5
nnmnkwii==0.0.21
numba==0.51.2
numpy==1.19.4
oauthlib==3.1.0
packaging==20.7
pandas==1.1.4
Pillow==8.0.1
playsound==1.2.2
pooch==1.3.0
prometheus-client==0.11.0
promise==2.3
protobuf==3.14.0
psutil==5.7.3
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
pypinyin==0.40.0
pyrsistent==0.17.3
PySocks==1.7.1
pysptk==0.1.18
pystoi==0.3.3
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==2.0.1
pytorch-wpe==0.0.0
pytz==2020.4
pyworld==0.2.12
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
schema==0.7.4
scikit-learn==0.23.2
scipy==1.5.4
sentencepiece==0.1.86
sentry-sdk==0.19.5
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smmap==3.0.4
SoundFile==0.10.3.post1
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.36.7
stempeg==0.1.8
style==1.1.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
threadpoolctl==2.1.0
toml==0.10.2
torch==1.8.0
torch-complex==0.2.0
torchvision==0.10.0
tqdm==4.54.0
typed-ast==1.4.1
typeguard==2.10.0
typing-extensions==3.7.4.3
Unidecode==1.1.1
update==0.0.1
urllib3==1.25.11
wandb==0.10.12
watchdog==1.0.1
websocket-client==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.0
